move history components [deployed]
