+++ src/scenes/Game/components/MoveHistory/components/Move/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import styled from '@emotion/styled';
 
-import { Color } from '../../../../../logic/constants';
+import { Color } from '../../../../../../logic/constants';
 
 const Wrapper = styled.div(
   ({ selected, piece }) => ({
