+++ src/scenes/Game/components/MoveHistory/index.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import styled from '@emotion/styled';
 import { Segment } from 'semantic-ui-react';
 
-import Move from './Move';
+import Move from './components/Move';
 import * as logic from '../../../../logic';
 import * as styles from '../../../../commonStyles';
